Fix handling of existing btrfs subvolumes.
Some of the highlights:

We weren't setting self.originalFormat correctly when implicitly
setting the btrfs format.

Also, we didn't have any code to discern subvolumes when parsing
/etc/fstab.

We were passing volume id as volume path/name when populating the
device tree.

Lastly, we were clobbering the subvol= mount option when mounting
a btrfs subvolume to check it for existing roots.