Provide a way to run apply methods even if a spoke wants us to skip elsewhere.
This first fixes a bug where a spoke's apply and execute methods were being run
even when skipTo was set.  In this case, we don't usually want apply to be run
because the user needs to do something else first.  In particular, skipping from
the dep solving error dialog (still in progress) to the software selection
screen was resulting in problems.

Having fixed that, some spokes do actually need to have apply run even when
skipping.  Storage is a prime example of that.