reboot after kickstart
Modify the continueButton to exit (allowing the anaconda exitHandler to
take control) if it is the last action. The button is expected to be
setup so that it is clear to the user what will happen (eg. REBOOT).

This also fixes the display of the quitDialog when rebooting.

Add a click of the final button when kickstarting with reboot or
shutdown.

halt also works, but requires a new version of pykickstart. Previously
the halt command was not working, it would set action to KS_SHUTDOWN and
there was no way for KS_WAIT to be set.