Add support for creating device based on a top-down specification.
General Approach:

Instead of an all-or-nothing approach, the goal is to come as close
as possible to satisfying the new device specification. Once the
container has been allocated, we simply adjust the new device's
size as needed to fit in the container.

This uses growable partition requests, but it then fixes their size
as soon as they have been allocated. It also changes doPartitioning
to reallocate container member devices each time instead of fixing
them on disk once the container is defined. Since the members are
not growable anymore once the container is defined, this is less
disruptive than it would otherwise be. It allows for maximum
flexibility to allocate the set of requests while still preserving
already-defined containers' sizes.

Notes:

Creation of md devices is completely untested.

Creation any device (md, lvm, btrfs) with striping, mirroring, or
any other RAID-like features is completely untested.

There is no support for container members of any type other than
normal partitions.