Add packaging log and scriptlet logging
This adds a new logfile for packaging (/tmp/packaging.log) which will
contain everything related to yum and rpm as well as other package
systems when they are added. This also re-adds logging of the rpm
scriptlet output.

The 'Installing ...' messages end up being logged twice. Because rpm
only logs to a file we capture the scriptlet output along with the
package install message so that you can associate scriptlet problems
with their package. This adds this file's log to our logs after the
transaction is complete (or fails).