Update rescue mode to continue working in the newui world.
This is still just the existing newt rescue mode interface, but with adding
advanced devices removed (requires other code that no longer exists).  Of
course, this all needs to be removed and replaced by something that does not
use newt in the future but for now, it's enough of a stop-gap to fulfill
release requirements.