dracut: fix inst.ks.sendmac (#826657)
Originally, parsing inst.ks.{sendmac,sendsn} happened in
parse-anaconda-kickstart.sh. But it turned out that sendmac needed to
happen after the modules were loaded, so commit 4fcc157 tried to fix
this by making it a function, and using initqueue to run that function
in the 'initqueue/settled' hook.

Well, it turns out initqueue doesn't like adding jobs that are function
names - the job ends up empty and nothing happens.

So instead, let's handle these two arguments directly in a script that
runs in initqueue/settled.