Add a text progress hub to do the install
This is a pretty simple hub that works with the doInstall function as a
background thread.  In the foreground we sit in a tight loop waiting for
input to the Q and then handle messages or progress bar updates.
Progress bar updates are handled by printing a pip (.) without a new
line to the screen.