dracut: suppress ks errors from missing %include
The kickstart parsing that happens in dracut is the "early" parsing - it
doesn't run %pre and it might not be connected to the network, so it's
possible we'll have %include lines that aren't yet valid.

KickstartParser has a 'missingIncludeIsFatal' attribute for this reason.
Set it to False (like loader did) to turn off these errors.