Make choosing an auto-selected page after refresh slightly less fallible.
Build a list, in decreasing order of preference, of pages that make sense
to initially have selected. If a page was selected at the time of the
refresh, it should probably be still selected afterwards. If that page is
no longer present, no problem. Move on to the next one in the list.