BTRFS magic for custom spoke.
For devices whose formatting is swap, remove BTRFS from the device type
combo.

Whenever the device type is switched to BTRFS, set fstype to btrfs and
lock it in.

Do not include btrfs as fstype unless device type is BTRFS.

Note: I made this a bit simpler by making BTRFS the last item in the
device type list and doing the same for btrfs in the fstype list.