Don't rely on selection staying selected when doing crazy things to it
LanguageSpoke inherits from the LanguageMixIn which is also used by the
WelcomeLanguageSpoke and contains the treeview with available languages and the
selection object. Obviously the selection doesn't stay selected when
WelcomeLanguageSpoke is hidden.
So in LanguageMixIn.initialize -- if there is nothing selected, try to use
self.data.lang.lang value first and if it is None, fallback to prefered_language
(English).

Fixes the issue with no translations and default language selected on the
SummaryHub before going to LanguageSpoke and back.