Add a gui password spoke
This spoke sets the root password, or if left blank disables the root
login.  Some level of validation is done and the user is trapped until
they get it right or press back twice on a weak password.

capslock monitoring is not enabled currently, but there is some stub
code commented out that deals with it.