Put traceback reports on a diet.
This updates what attributes get written out to the traceback file for the
newui.  In particular, it avoids writing out all the yum stuff that makes the
traceback file grow over 2 MB and take nearly a minute to pickle.

It's entirely possible this is either skipping too much or writing out too
much still, but we can work on that over time.