parse-kickstart: handle 'network --ipv6=auto ...'
This is kind of silly, since IPV6_AUTOCONF=yes is the default (according
to sysconfig.txt from 'initscripts'). But it might not always be the
default, so let's just make sure we do what the user wants.

Also, it's not our job to set policy; if DHCPV6C=yes should disable
IPV6_AUTOCONF that's up to initscripts/NetworkManager. So leave it alone
otherwise.