Setup python path /after/ we've done updates
If we try to setup the python path before /tmp/updates/ exists, the path
will get thrown out and won't be used even if that path will get created
at some point later.  So just re-order things.

This is ugly, and shouls only live for the F18 Alpha.  After that we
should be using /run/install/updates directly and thus the path can be
set there early.