Explicitly disable the rootpw lock (#853788)
Turns out that while pykickstart objects will default to the lock being
False, there is an interactive-defaults.ks that is being loaded by
anaconda, and in /that/ file the root is default to --lock.  Since we're
setting a password, it makes sense to ensure that the account is not
locked.