Only check media if we really want it (#853404)
This provides a way to opt-out of the check for existing mounted media.
If a user has gone into the source spoke and picked a mirror, we
shouldn't ignore that choice by checking for valid mounted media.

In the case of a network stage2 that has a valid repo, we'll find that
when the hub / spokes are initialized.  If a user visits the source
spoke and just clicks back (leaving autodetect selected, but no media)
we just return from apply() so that initial discovery of working stage2
should remain.