Add error handling around significant ui-initiated storage operations.
If automatic partitioning fails, we reset the storage instance to its
prior state. The custom spoke operates on a copy of the storage
instance, so that means it gets reset to whatever the state was upon
entering the custom spoke. For the storage spoke we have little choice
but to actually call self.storage.reset() which actually re-scans the
devices.

For other operations within the custom spoke, the goal is to restore
things to the state prior to the failed action. If that fails, the
custom spoke will simply reset its storage instance as described above
for automatic partitioning failures.