Handle newDevice partitions smaller than the default of 500MB. (#853125)
With larger partitions we were starting with the default size of
500MB and growing as close as possible to the requested size, but
for smaller requests we were attempting a fixed-size allocation.
That's no good because it's a change in behavior for requests
below a somewhat arbitrary size and the difference is the smaller
(fixed) request allocations are more likely to fail. The solution
is to always use a base size of 1MB instead of 500MB. It works
without messing up partition growth because all partition sizes
get fixed after their first allocation, which means that at any
point where we're growing a request with a base size of 1MB we
know that any other growable requests will have the same base size.