Add a way to test exception handling
Instead of adding command line option throwing exception in one place
of the code, let's use SIGURS1 as signal for raising exception in anaconda.
This way, it is possible to raise exception in different phases of the
installation and we can add code to prepare worst case scenario testing
as much resolved bugs in exception handling as possible.