Improve growth check when deciding where to allocate new partitions.
In the case where the new partition is going to be the first logical
partition on the disk (and therefore require creation of an extended
partition) we were failing to correctly set things up to perform the
growth calculation. This led to sub-optimal placement of partitions.

This also fixes a small bug that caused us to jump to the new
candidate free region even when it offered the same growth
potential as the old/current one.