Rescan all disks when moving back through UI (#804557)
When the user has selected a clearPartType it can interfere with the
discovery of partitions to be upgraded when moving back through the UI
to sections before clearPartType selection, or to before the upgrade
screen.

This fixes those cases by adding a new option to storage.reset, named
examine_all, that when set to True will cause it to reexamine all disks
regardless of the clearPartType setting.

Resolves: rhbz#804557