Handle automated installs
This follows the same design as GUI.  If the install is automated, and
all the spokes are complete, proceed into the installation phase.
However if there are incomplete spokes, let the user finish those
spokes.  As soon as all spokes are complete jump straight to the install
phase.

Note that any new spoke that has a background process that could make it
not be "ready" immediately is going to have to be accounted for here,
which is ugly.  Perhaps a object attribute that is a list of potential
threads to join would work?