Don't use grey for the status text of a SpokeSelector (#855638).
By hardcoding a color, it makes the status text almost invisible on the KDE
live disc where the theme is darker to begin with.  Instead, use the theme
colors for everything.

This isn't a perfect fix:  If you focus a SpokeSelector, the text will still
be black instead of white like everything else.  However, I've been unable to
make that work.