Add a guard for testing if we can modify runtime system
Using "if flags.liveCDinstall or flags.imageInstall or flags.testing" over
and over is incovinient, one may forget to involve some of the conditions
and there is no logging of the skipped actions.

This patch adds flags.can_touch_runtime_system function, that can be used as
a guard and that logs skipped actions. It can also be easily mocked or
masked for the needs of testing.