Hide the ISO install source if you've nuked all possible drives (#858088).
This is a little touchy, but you first have to go to the source spoke.  This
will cause all widgets to be displayed.  Then go to the destination spoke and
choose to autopart all your drives.  Then go back to the source spoke.  We need
to both set the ISO-related widgets to no_show_all and hide them due to the
previous visit.