In the install options dialogs, call out how much space is on selected disks.
The idea here is that the dialog says you can make enough space to install by
resizing or deleting, but only ever mentions how much space is available if you
do resizing.  That seems weird to me, so I've added another row that shows how
much space is in all the selected disks.