Do not use constant value in SoftwareSpoke's completed property
packagesSeen is a value, that should not be changed in spokes. Using
it as a condition for the completed property results in no way of
completing software spoke in kickstart installation without %packages
section. Without %packages section in kickstart, SoftwareSpoke's
completed method should use the same code as in manual installations.