Add in a locale mapping to avoid incorrect system settings (#858591).
This is pretty terrible, sorry.

Basically the problem is that we need to write out full locale settings to
/etc/sysconfig/i18n or lots and lots of things on the system are broken.
For instance, "de_DE.UTF-8" is good but "de.UTF-8" is not.

Unfortunately I can find no suitable way to extract this information via
python-babel.  I think our intention is that using geoip will get us the
full locale, but there's no time to do that now.  lang-table used to contain
this mapping for us, but I really don't want to add that file back.  This
fix should be considered temporary.