Work with VConsole keymap and X layouts separately
We need to set VConsole keymap and X layouts separately. We may
get them all from kickstart, but if not, we have to convert the
given one to the missing one and save the configuration. If X11
configuration is written before the server is started, it is
loaded. This way we can get rid of using XklWrapper in main anaconda
script.

Also we need to write one more configuration file. And it's better
to try to write all configuration files ignoring errors and in case
some error(s) appeared, raise exception at the end of the procedure.

Resolves: rhbz#853877
Resolves: rhbz#856362
Resolves: rhbz#859867