Implementing a delay algorithm for slow bios storage devices
When anaconda is booted with ks=hd:<dev>:<path>  there is an algorithm  to
retry up to 5 times or 31 seconds for some USB thumb drives and hard drives
which are initialized slow. Such algorithm is not available in the code path
when booted with the option "ks=bd:device:/path/to/ks.cfg".

Resolves: rhbz#848741