Add ARM-OMAP class to create a uboot partition to support the boot-loader.
Break out the setup of ARM-OMAP systems, as they need to have a vfat partition
at the start of the disk in order for the SOC to load the initial boot code.
This also sets the weight for the '/' partition to ensure it is the last
partition on the image.

Signed-off-by: Brian C. Lane <bcl@brianlane.com>