Add a service to run anaconda directly on the tty
When installing on s390x, we just want to run pure anaconda on the tty.
This will either detect s390x and print out ssh details for the user to
ssh in to continue the install (and get tmux), or if a kickstart was
provided and the RUNKS option was used it will just blast through the
kickstart in cmdline mode.  Tmux doesn't work in x3270, the terminal for
s390x, so we have to avoid using it.