Revive reipl (#860244)
reIPL used to be ran via dispatch, and would trigger either a reboot or
a shut down.  Now we'll run it as part of the s390 bootloader install.
The message warning users to reipl themselves is somewhat lost here,
something for another day perhaps.  This at least gets the system
working right.