The environment window needs a vertical scroll bar (#865066).
With the addition of Cinnamon and MATE and some others, there's now entirely
too much stuff in the environment window to fit without scrolling.  Plus, I
guess smaller screens always needed to scroll.  Without a scroll bar, GTK
puts this window in a really bizarre place.  See the screenshot attached to
the bug.