Fix graphical kickstart with %packages data
Without this change we were automatically selecting the first
environment from the list, and adding it to the existing package data.
This means no matter what you had in %packages you also got the GNOME
environment.  Obviously this is wrong, so we don't autoselect if we're
doing an automated install and have package data.