Remove the code for removing an entire Root all at once.
First, it simply does not work right now.  Secondly, it's not at all obvious
in the UI how you would go about doing this.  Click on the expander and then
hit remove?  Yes, but then the expander collapses and you still have a
mountpoint displayed on the right.  Which thing are you deleting then?

We need a way to do this, but I don't think this code is it.