Avoid using mount --move on shared paths (#853508)
move doesn't work because /run/ is a shared filesystem.  So do as
systemd suggests, mark the filesystem as private.  We don't know of any
reason in anaconda for the filesystem to be shared anyway.

This hack can go away if/when the kernel ever allows moving mounts
within a shared filesystem.