Configure new-kernel-pkg to keep tboot configuration on updates (#742885)
If we want tboot to keep getting used, we need to configure
new-kernel-pkg to tell grubby to do that.  When grubby finds a  multiboot
stanza as the default template, it won't use the multiboot parts of the
template unless specifically configured to.  This patch adds a
configuration option to tell grubby that we always want to use the tboot
multiboot kernel by default when we add a new kernel.

Signed-off-by: Peter Jones <pjones@redhat.com>