Hook up the "Remove Packages" button on the dep solving error screen.
All this does is set skip_broken and yum does all the rest.  Note that we
are still doing dep solving a second time, when the progress hub is up, but
adding this setting to the Payload object means skip_broken will be in effect
there too.  We still need proper error handling for other errors that might
occur at that point.

This should take care of #865073.  I can't think of any cases where you can
get past the first hub with a dep solving error.