Set the status text in the SpokeSelector widget differently now.
This is required by the previous patch.  We are storing pango markup in the
language store now, and pango does not appear to support nested spans.  So,
we need to set attrs instead.  Also I've condensed things a bit here to make
sure the code is only in one place.