check disklabels when calculating free space (#863892)
You need at least one of the platform's disklabels in order for the free
space on the disk to be useful. If none of the disks labels will work we
set the free space to 0 and allow the resize dialog to be used to remove
all the partitions which will relabel the disk.