Fix a race condition with kickstarts (#868834)
We were simluating the continue click before consuming all the messages
in the queue, which would cause the next hub to be in use, but we were
attempting to process messages for the previous hub.  It would cause a
crash of anaconda, but not a backtrace that could be caught with our
autoreporter.  The solution is to only continue when the queue is empty.

This also makes use of a property to see if we are ready to continue
instead of duplicating code.