Apparently necessary kpartx changes (#867593)
Don't always force kpartx to use a delimiter. The problem is that there
are all these places that could be triggering creation of the partitions
and they need to all use the same rules for when to use a delimiter
between disk and partition number. Otherwise, how can parted know when
to use a delimiter when adding a partition to a dm disk? You'd think the
answer would be "always use a delimiter" or "never use a delimiter" but
that's not how the world works.

Also, pass -s to kpartx to wait for devices to be created before
returning since udev should be doing all the work nowadays.