Fix nfsiso repo selection (#871648)
Two fixes here, one set the path to INSTALL_TREE because we haven't
moved anything over to ISO_DIR yet.  Two account for the isys mount
count for paths.  When we move a mount point the isys count must be
updated so that an attempt later to mount the iso into the INSTALL_TREE
won't be short circuited.

One could create an (_)isys call that does the move and does the
accounting, but that would add more C code, which is an anti-goal.