Move the custom partitioning's Apply Changes button.
Right now it gets lost floating around in all the free space in the bottom
right side of the screen, or even worse gets confused with the Finish
Partitioning button.  This moves it to the left and up right under the
customize expander where it will hopefully be easier to find.