Mark strings at the top of spokes with N_; translate later with _ (#872791).
Spokes are imported very early (when the UI is being constructed), which is
before a language has been chosen.  Thus, any strings at the top of the file
will be set up with the original translation, not the translation the user
selects.  The fix is to simply translate them when they are needed and not
before.