Correctly handle toggle of encryption state for devices. (#873445)
Don't even allow toggling of encryption for existing devices that have
their encryption applied at some lower level (eg: what our autopart
produces).

If the encryption is applied to container members (lvm, btrfs), let
Storage.setContainerMembers handle the changes. If toggling encryption
for an existing device or for any device for which we apply encryption
to the leaves (md, partition), handle the change of state directly in
_save_right_side.