Account for autopart swap size when checking free space.
Only include the swap request size in the case where we're about to
tell them they can run autopart in the existing free space, since
that's the main place it can be a surprise to find that you do not
actually have space for the selected software.

For the bug referenced below we still need to redo the space check
each time a new software selection is applied.

Related: rhbz#853636