Add support for preexisting whole-disk formatting. (#870476)
In the resize/reclaim dialog you can trigger creation of a disklabel
on a disk containing whole-disk formatting by selecting "Delete".

In custom, you can set mountpoint, reformat, encrypt. If you want to
remove the filesystem and set the disk up with a disklabel, hit the
minus/remove button while the whole-disk filesystem is selected.