Don't let mount/umount block python threads (#873600).
Without these lines, the IO blocks everything in python until complete, which
means setting up a bad NFS mount in the source spoke gives you an unresponsive
hub until the NFS timeout hits.

This is only a patch.  The better fix would be to use subprocess for mounting
instead of our own C blob.  The only reason this C blob exists is because
loader also required it, and no one's found the time to convert this old code.