Vastly simplify the process for applying changes from custom spoke.
I hit a bug triggered by repeatedly toggling encryption for a vg
while also doing some size changes that revealed a problem in the
current apply/propagate_actions approach. There's just too much
that can happen to try to duplicate/cover it all.

Just replacing the main Storage instance isn't an option since each
of the ui spokes holds a reference to it. However, all of the data
is really in just a few data attributes of Storage and DeviceTree.