Adjust right margin for MountpointSelector (mizmo).
The mountpoint selector has a problem where all mountpoints except the
selected one are flush with 0 margin against the right wall of the box
they're in, and the selected mountpoint gets 6 px of margin between it
and the wall.

This patch makes that margin more consistent, so both selected and
unselected mountpoints in the mountpoint selector list on the left hand
side of the custom partitioning spoke get a consistent 12 px margin.