Use the main loop to control displaying the resize dialog.
Emitting signals from the function that handles other signals is dangerous
given the synchronous nature of GTK signals, so this was never a good idea
to start with.  Then, displaying the dialog from the main loop means it
gets displayed in the lightbox correctly instead of over a completely grey
background.