Correct colors for selected items in mountpoint selector widget (mizmo).
This patch adjusts the coloring for selected mountpoints in the
mountpoint selector widget. Previously, some of the text labels appeared
grey or black on the blue selected background, making it harder to read
them. This patch adjusts those so they go white when selected, and black
or gray (depending on the label) when not selected.