update mdraid superBlock space calculation (#876450)
mdadm 3.2.4 made a major change in the amount of space used for 1.1 and
1.2 metadata in order to reserve space for reshaping. See commit
508a7f16 in the upstream mdadm repository.

This has resulted in a large difference between what anaconda thinks the
size of the array should be and what it actually is.

We now use 2.0MB for metadata that isn't 1.1 or 1.2
For 1.1 and 1.2 we calculate the reserved space value and use that.

Resolves: rhbz#876450