Fix code to lock encryption checkbutton for LV in existing VG (#877871)
The code to decide when to lock the encryption checkbutton was too
simple. It was possible to toggle it for existing LVs that had been
marked for reformat previously. It was also possible to toggle it for
new LVs in existing VGs whose PVs were encrypted. It should be active
and insensitive whenever there is encryption on existing devices
anywhere in the stack since we cannot change that.