Rework actions in the resize dialog to avoid shortcomings (#866209, #867770).
There are three problems with the current resize action UI:

(1) You can't discover that there's a drop down in the action column because
there's no down arrow.  GTK simply doesn't do that.

(2) We can't filter out shrink from the drop down on a per-filesystem basis
without layering on all the usual filter store crud.

(3) Something about interacting with a GtkCellRendererCombo causes the layering
of dialogs to get screwed up such that coming back to the storage spoke after
visiting the resize dialog appears locked up, because the dialog is displayed
under the spoke.

This commit fixes all these problems by getting rid of the drop downs in the
action column and adding a communal set of tool buttons for setting the action
you want.