Add checkbox for setting language default layout (#866887)
Some users expect that when they select language Anaconda automatically
sets their keyboard layout to the one that is default for the selected
language. But this would not be a good implicit behaviour because we would
be changing layout in background without letting user now. Adding checkbox
for the behaviour some users want seems to be a best solution.