Remove partitions from all appropriate DiskLabel instances. (#870586)
Now that StorageDevice's format and originalFormat attributes are
distinct instances of a DeviceFormat class we have to explicitly
duplicate partition removals when we aren't reinitializing the disklabel.
All removals are currently done on originalFormat and format still has
all of its original partitions, which becomes a problem when we try
to add new ones to it.