Do not list some layouts twice (#882526)
If some layout is related to two (or more) languages, we get it twice
while iterating over languages and their related layouts. Since we use
a mapping from a layout name (e.g. 'us') to decription ('English (English (US))')
the result is that some items are listed twice in the list of all available
layouts. This needs some further fixing, but I'm leaving this for the related
post-F18 UI changes (see #859465 for more information).