Add/remove the HDISO source from protectedDevSpecs (#882147).
Should storage ever get reset (which it can in the case of moving back to
the hub from reclaim, resulting in an error), any protected attributes will
be wiped.  protectedDevSpecs is the definitive source for what's protected
and what's not.

The dev.protected lines are still there in case we do not hit an error.