Handle partition removals regardless of deepcopy. (#884896)
A StorageDevice containing a DiskLabel will have both an originalFormat
and a format attribute, both of which are DiskLabel instances. They are
copied using a shallow copy, so they share _partedDisk and _origPartedDisk
attributes. That means you only have to do partition removals from the
parted Disk once.

When you make a deep copy of such a StorageDevice, things change. The
originalFormat and format no longer share a single _partedDisk or
_origPartedDisk. In this case, it is sometimes necessary to remove
a partition from both self.disk.format and self.disk.originalFormat.