Add a gradient background to spoke headers (mizmo, clumens).
The idea here is that by adding a gradient up top, the "Done" button looks
more like it belongs with the header than with the widgets down in the
spoke.  It also subtly makes the spoke look like a subpart of the hub.

Adding this required adding an additional internal widget, which means also
changing all existing glade files to include this internal widget.  Make sure
to build new glade support and install locally (or use the proper makefile
targets) before editing.