Add handling for incomplete lvm/md devices. (#876441)
The strategy is to keep the vg/md in the devicetree, but to not include
them in devicetree.devices or the results of getDeviceBy{Name,Path}.
In the custom spoke we display a special options page for them with a
brief explanation and the option to remove the device or leave it be.

When removing an incomplete vg we don't actually do anything to remove
the vg since lvm gives us no way to specify one of several vgs with the
same name. All we do is wipe the pvs that we have access to.
(cherry picked from commit 531136b3eaf3dca07daa2f4a50fb8a5c980ef7fb)