Try fallback if none exactly matching language is found (#891487)
From the language selection we get e.g. "Greek" as a language name. But
while iterating over all languages libxklavier knows we get "Greek, Modern
(1453-)" which is not an exact match. If exact match is not found we should
try to find language starting with the string we got from the language
selection.