start vnc without ip address (#832510)
In some cases the vnc server may start before the system has an IP
address. This isn't a fatal problem, it just makes it harder for the
user to know where to connect their vnc client.

This adds a 5 second retry when getting the IP. If that fails it will
still start the vnc server, and will print out instructions for
how to manually discover the IP address.