Remove unused functions and move storage-specific utils to storage.
This is the first pass at setting things up to split storage out into
a separate package.

It replaces the guts of iutil.execWithRedirect and iutil.execWithCapture
with a single backend function, _run_program. The main thing I can think
of that might be cause for complaint is that _run_program does not log
each output line as the external program outputs it, so you get the whole
output buffer at once. Storage gets its own copy of _run_program and
all of the calls in storage use the copy from storage.util.

The reason I moved all of the iutil arch functions into storage is that
they are mostly used in storage and I didn't want there to be two copies
of the whole set.

The rest is removing of unused things and moving of things only used by
storage into storage.