Check country_layouts is not None when using it (#893026)
In get_default_lang_country_layout we may have some layouts specified
for the given language and none for the given country. In such case
we should just return the first matching the given language. And vice
versa if there is no layout for the given language but there is some for
the given country.