On storage, remove the "Continue" button and make "Done" do it all (#882737).
This is another attempt to fix some of the confusion of having both a Continue
and a Done button on the same screen.  To clear it up, I've completely removed
the Continue button and made the Done button do everything that Continue used
to do.  This makes the storage spoke a little more standalone instead of the
first screen in a series.