Start laying groundwork for splitting storage out of pyanaconda.
Remove platform's use of bootloader in preparation for moving platform
into storage, since that's the only code that uses it.

Move some logging initialization code in network.py into a function so
it does not run automatically on import.

Add some code to storage to control whether or not it is being used by
anaconda versus as a normal python module/library.