Rework the reclaim dialog to have a resize slider.
This required changing the toolbar of buttons into a regular hbox.  I've also
removed the percent bar since that's duplicate information and gets rid of a
couple numbers (this dialog has way too many), and changed the wording in the
reclaimable space column for unshrinkable devices to say so.  That gets rid of
some more numbers.

Related: #875944