Display the background gradient image from a map signal handler.
Doing it from the init method means we need to call gtk_widget_realize so all
the gdk and cairo stuff has something to operate on.  However, that realize
call screws up glade so that the window gets displayed separately and widgets
cannot be clicked upon.

This reorganization moves the gradient setup to after the widget has been
realized.