Add error reporting to threadMgr
We need to be able to tell when a thread has quit because of an error.
This adds the ability to save thread traceback information in threadMgr
and a new method, .wait() that can be used instead of join. It will
raise an error if the thread quits because of an Exception.