Refactor and cleanup exception handling pieces
Before this patch our UserInterface classes somehow "implicitly"
inherited python-meh's *Intf classes by providing two methods with right
names. This patch adds a meh_interface property to our UserInterface
classes that returns an instance of a class that really is inherited
from the python-meh's *Intf class or actually is an instance of the
python-meh's *Intf class in case of gui and tui respectively. It also
adds a tty_num property to our *UserInterface classes that can be used
in runDebug to get user to the right console before and after the pdb
session automatically.