Add the advanced storage UI and hook it up.
This large patch adds the new filter spoke off the storage spoke, consisting
of all the UI needed for searching and multipath, plus the code to make those
two work and the framework necessary for making the other tabs work.  I've
added some preliminary zSeries stuff, but have not yet tested it.