Run exception handling in the main thread also in TUI
We need to get exception handling to the main thread in both graphical
and textual user interface. This patch adds the code necessary to get
exception handling to the main thread in text mode by using the async
handling based on the communication queue. We just need to distinguish
between exception in the main thread or in a non-main thread.

If the exception appears in a non-main thread a message with the
exception data is sent to the queue instead of running the exception
handling. When such message is handled (which happens in the main
thread), the exception handler is run again and this time it runs the
exception handling.

We need to tell python-meh to use our raw_input because the standard one can be
called only once (the other calls are queued) and thus the one called by the TUI
will block the python-meh's one. And then the same applies to Pdb session run
from the python-meh and so on.