Bring back the askmethod parameter (#889887).
We've got this shiny new method selection UI but if you want to use it, you've
first got to sit through the delay of fetching metadata for the closest mirror
which might be completely wrong for you.  Thus, this parameter will skip the
default action.  Entering and leaving the source spoke will unset this
parameter.