Try to import modules the standard way first in collect
This ensures that the correct package structure is present
in sys.path for modules and addons that make use of it.

The package-less addons will be imported file by file with
the package structure ensured by dummy empty modules.

This also changes module masks for addons to make the change
work properly.